requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
matplotlib==3.7.2
scipy==1.11.3
scikit-learn==1.3.0
scikit-image==0.21.0
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
joblib>=1.3.0
